Refactor @Preview methods to separate files
Moved @Preview methods from component source files to dedicated `Preview.kt` files to improve code organization and separation of concerns. This includes:
- `ThemeItem.kt` -> `ThemeItemPreview.kt`
- `DownloadUnreadBadge.kt` -> `DownloadUnreadBadgePreview.kt`
- `HeaderCard.kt` -> `HeaderCardPreview.kt`
- `InfoWidget.kt` -> `InfoWidgetPreview.kt`
- `TextPreferenceWidget.kt` -> `TextPreferenceWidgetPreview.kt`
- `SwitchPreferenceWidget.kt` -> `SwitchPreferenceWidgetPreview.kt`
- `EmptyScreen.kt` -> `EmptyScreenPreview.kt`
- `InfoScreen.kt` -> `InfoScreenPreview.kt`

Verified compilation with `./gradlew app:compileStandardDebugKotlin`.
