+++ app/src/main/java/org/nekomanga/presentation/components/DownloadUnreadBadge.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -20,7 +19,6 @@ import androidx.compose.ui.graphics.Outline
 import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
@@ -281,46 +279,4 @@ private class SlashedRoundedShape(
                 }
         )
     }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreview() {
-    Surface { DownloadUnreadBadge(true, true, 100, true, 100, 0.dp) }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreview10() {
-    Surface { DownloadUnreadBadge(true, true, 10, true, 1, 0.dp) }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreviewSmall() {
-    Surface { DownloadUnreadBadge(true, true, 1, true, 1, 0.dp) }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreview2() {
-    Surface { DownloadUnreadBadge(false, true, 100, true, 100, 0.dp) }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreview3() {
-    Surface { DownloadUnreadBadge(true, false, 100, true, 100, 0.dp) }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreview4() {
-    Surface { DownloadUnreadBadge(false, true, 100, false, 100, 0.dp) }
-}
-
-@Preview(showBackground = true, apiLevel = 32)
-@Composable
-private fun DownloadUnreadBadgePreview5() {
-    Surface { DownloadUnreadBadge(false, true, 1, false, 100, 0.dp) }
 }
