+++ app/src/main/java/org/nekomanga/presentation/components/HeaderCard.kt
@@ -1,9 +1,7 @@
 package org.nekomanga.presentation.components
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ElevatedCard
@@ -12,7 +10,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.nekomanga.presentation.theme.Shapes
 import org.nekomanga.presentation.theme.Size
@@ -38,12 +35,4 @@ fun DefaultHeaderText(text: String) {
         color = MaterialTheme.colorScheme.onSecondary,
         modifier = Modifier.fillMaxWidth().padding(12.dp),
     )
-}
-
-@Preview
-@Composable
-private fun HeaderCardPreview() {
-    Box(modifier = Modifier.statusBarsPadding()) {
-        HeaderCard { DefaultHeaderText(text = "My Test Header") }
-    }
 }
