+++ app/src/main/java/org/nekomanga/presentation/components/theme/ThemeItem.kt
@@ -32,7 +32,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import eu.kanade.tachiyomi.util.system.isInNightMode
 import org.nekomanga.R
@@ -201,10 +200,4 @@ fun AppThemePreviewItem(
             }
         }
     }
-}
-
-@Preview
-@Composable
-private fun PreviewThemeItem() {
-    Surface() { ThemeItem(theme = Themes.Pink, isDarkTheme = false, selected = false) {} }
 }
