+++ app/src/main/java/org/nekomanga/presentation/screens/settings/widgets/InfoWidget.kt
@@ -7,18 +7,13 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Info
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.PreviewLightDark
 import jp.wasabeef.gap.Gap
-import org.nekomanga.R
 import org.nekomanga.presentation.components.NekoColors
-import org.nekomanga.presentation.theme.NekoTheme
 import org.nekomanga.presentation.theme.Size
 
 @Composable
@@ -36,10 +31,4 @@ internal fun InfoWidget(text: String) {
             style = MaterialTheme.typography.bodyMedium,
         )
     }
-}
-
-@PreviewLightDark
-@Composable
-private fun InfoWidgetPreview() {
-    NekoTheme { Surface { InfoWidget(text = stringResource(R.string.download_ahead_info)) } }
 }
