+++ app/src/main/java/org/nekomanga/presentation/screens/settings/widgets/TextPreferenceWidget.kt
@@ -2,21 +2,16 @@ package org.nekomanga.presentation.screens.settings.widgets
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Preview
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.tooling.preview.PreviewLightDark
 import org.nekomanga.presentation.components.NekoColors
 import org.nekomanga.presentation.screens.settings.BasePreferenceWidget
-import org.nekomanga.presentation.theme.NekoTheme
 import org.nekomanga.presentation.theme.Size
 
 @Composable
@@ -67,26 +62,4 @@ fun TextPreferenceWidget(
         onClick = onPreferenceClick,
         widget = widget,
     )
-}
-
-@PreviewLightDark
-@Composable
-private fun TextPreferenceWidgetPreview() {
-    NekoTheme {
-        Surface {
-            Column {
-                TextPreferenceWidget(
-                    title = "Text preference with icon",
-                    subtitle = "Text preference summary",
-                    icon = Icons.Filled.Preview,
-                    onPreferenceClick = {},
-                )
-                TextPreferenceWidget(
-                    title = "Text preference",
-                    subtitle = "Text preference summary",
-                    onPreferenceClick = {},
-                )
-            }
-        }
-    }
 }
