+++ app/src/main/java/org/nekomanga/presentation/screens/settings/widgets/TextPreferenceWidgetPreview.kt
@@ -1,0 +1,31 @@
+package org.nekomanga.presentation.screens.settings.widgets
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Preview
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.PreviewLightDark
+import org.nekomanga.presentation.theme.NekoTheme
+
+@PreviewLightDark
+@Composable
+private fun TextPreferenceWidgetPreview() {
+    NekoTheme {
+        Surface {
+            Column {
+                TextPreferenceWidget(
+                    title = "Text preference with icon",
+                    subtitle = "Text preference summary",
+                    icon = Icons.Filled.Preview,
+                    onPreferenceClick = {},
+                )
+                TextPreferenceWidget(
+                    title = "Text preference",
+                    subtitle = "Text preference summary",
+                    onPreferenceClick = {},
+                )
+            }
+        }
+    }
+}
